+++ engine/src/main/java/dev/gamekit/utils/Timeout.java
@@ -9,7 +9,8 @@ public class Timeout {
   Task task;
 
   public Timeout(long duration, Task task) {
-    if (duration < 0) throw new RuntimeException("Timeout duration cannot be negative");
+    if (duration < 0)
+      throw new RuntimeException("Timeout duration cannot be negative");
     if (task == null) throw new RuntimeException("Timeout task cannot be null");
     this.completed = duration == 0;
     this.duration = duration;
@@ -24,13 +25,14 @@ public class Timeout {
    * task's {@link Task#run() run()} method is executed.
    */
   public void onUpdate() {
-    if (!completed) {
-      duration = java.lang.Math.max(0, duration - Application.FRAME_TIME);
+    if (!completed)
+      duration = java.lang.Math.max(
+        0, duration - Application.FRAME_TIME
+      );
 
-      if (duration == 0) {
-        completed = true;
-        task.run();
-      }
+    if (duration == 0) {
+      completed = true;
+      task.run();
     }
   }
 }